examples/imagenet-classifier-resnet/requirements.txt
requests
torchvision
Flask
numpy
deepmux
Pillow
